+++ packages/vue/src/browser/resolver/index.ts
@@ -1,0 +1,2 @@
+export * from './resolver-protocol';
+export * from './path-resolver';