+++ packages/vue/src/package.spec.ts
@@ -1,0 +1,4 @@
+describe('vue package', () => {
+
+    it('support code coverage statistics', () => true);
+});